Use SmsMfa component for MfaForm, ResetPasswordForm, and VerifyForm
